+++ include/heimdall/matrix.h
@@ -11,15 +11,15 @@ HEIMDALL_NAMESPACE_BEGIN
 class Matrix {
   public:
 	/// Matrix public data
-	double m[4][4];
+	float m[4][4];
 
 	/// Matrix public method declarations
 	Matrix();
-	Matrix(double _m[4][4]);
-	Matrix(double t00, double t01, double t02, double t03,
-			double t10, double t11, double t12, double t13,
-			double t20, double t21, double t22, double t23,
-			double t30, double t31, double t32, double t33);
+	Matrix(float _m[4][4]);
+	Matrix(float t00, float t01, float t02, float t03,
+			float t10, float t11, float t12, float t13,
+			float t20, float t21, float t22, float t23,
+			float t30, float t31, float t32, float t33);
 
 	bool operator==(const Matrix& mat) const;
 	bool operator!=(const Matrix& mat) const;
